pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "once"
dynamic = ["version"]

dependencies = [
    "numpy==1.26.4",
    "opencv-python-headless==4.10.0.84",
    "pillow==11.0.0",
    "pyarrow==18.0.0",
    "pyproject_hooks==1.2.0",
    "rerun-sdk==0.19.0",
    "scipy==1.14.1",
    "typing_extensions==4.12.2",
]

description = "Development kit for ONCE dataset"
readme = "README.md"
requires-python = ">=3.8"
keywords = [
    "dataset",
    "machine-learning",
    "autonomous-driving",
    "point-cloud",
    "images",
    "lidar",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",

]
authors = [
    { name = "Yin Guobing", email = "yinguobing@gmail.com" },
    { name = "m-Just", email = "mjust.lkc@gmail.com" },
    { name = "nmzfrank" },
]

[project.urls]
Homepage = "https://github.com/yinguobing/once_devkit"
Repository = "https://github.com/yinguobing/once_devkit.git"
Issues = "https://github.com/yinguobing/once_devkit/issues"


[tool.hatch.version]
path = "src/once/__about__.py"
